Trying to make the logging suck less. Before info() was a wrapper around a condition check for log level and then console.log.apply(). The Problem with this was that the console output would always say that the console.log message line number was withing info() not where info() was called.
Now info(), warn(), & crit() take no arguments and retern a boolean about whether
or not we are at the given level for logging. So we call:

  info() && console.log(...)

and info() is just a guard on whether or not to call console.log(...).

It takes more horizontal characters (hence more line wrapping), but the log
line number reporting is accurate.

Like I've said this whole scheme sucks.